CRUD user

+++ src/main/java/vn/hoidanit/jobhunter/service/UserService.java
@@ -4,12 +4,18 @@ package vn.hoidanit.jobhunter.service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import vn.hoidanit.jobhunter.domain.ResultPaginationDTO;
 import vn.hoidanit.jobhunter.domain.User;
 import vn.hoidanit.jobhunter.domain.dto.Meta;
+import vn.hoidanit.jobhunter.domain.dto.UserResponseDTO;
+import vn.hoidanit.jobhunter.domain.dto.user.UserDTO;
+import vn.hoidanit.jobhunter.domain.dto.user.UserMapper;
 import vn.hoidanit.jobhunter.repository.UserRepository;
+import vn.hoidanit.jobhunter.util.constant.GenderEnum;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,28 +24,24 @@ import java.util.Optional;
 public class UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final UserMapper userMapper;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, UserMapper userMapper) {
         this.passwordEncoder = passwordEncoder;
         this.userRepository = userRepository;
+        this.userMapper = userMapper;
     }
 
-    public User handleCreateUser(User user) {
-        User isUserExist = this.userRepository.findUserByEmail(user.getEmail());
-//        if (isUserExist.isPresent()) {
-//            throw new ResponseStatusException(
-//                    HttpStatus.BAD_REQUEST,
-//                    "The user with email " + user.getEmail() + " already exists."
-//            );
-//        } else {
-        String rawPassword = user.getPassword();
-        String hashed = this.passwordEncoder.encode(rawPassword);
+    public User handleFindUserById(long id) {
+        Optional<User> optionalUser = this.userRepository.findById(id);
+        if (optionalUser.isPresent()) {
+            return optionalUser.get();
+        }
+        return null;
+    }
 
-        user.setPassword(hashed);
-        return this.userRepository.save(user);
-
-//        }
-
+    public User handleFindUserByEmail(String email) {
+        return userRepository.findUserByEmail(email).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Không tìm thấy email: " + email));
     }
 
     public ResultPaginationDTO handleGetAllUser(Specification<User> spec, Pageable pageable) {
@@ -60,36 +62,55 @@ public class UserService {
         return rs;
     }
 
-    public User handleGetUserById(long id
-    ) {
-        Optional<User> optionalUser = this.userRepository.findById(id);
-        if (optionalUser.isPresent()) {
-            return optionalUser.get();
+    public UserResponseDTO handleCreateUser(UserDTO.Create userDto) {
+        if (userRepository.existsByEmail(userDto.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Email đã tồn tại");
         }
-        return null;
 
+        String rawPassword = userDto.getPassword();
+        String hashed = this.passwordEncoder.encode(rawPassword);
+
+        userDto.setPassword(hashed);
+
+        User entity = userMapper.toEntity(userDto);
+
+        User saved = userRepository.save(entity);
+
+        UserResponseDTO resp = userMapper.toUserResponseDTO(saved);
+
+        return resp;
     }
 
-    public User handleUpdateUser(User reqData) {
-        User currentUser = this.handleGetUserById(reqData.getId());
+    public UserResponseDTO handleUpdateUser(UserDTO.Update reqData) {
+        User existing = this.handleFindUserById(reqData.getId());
 
-        if (currentUser != null) {
-            currentUser.setEmail(reqData.getEmail());
-            currentUser.setName(reqData.getName());
-            currentUser.setPassword(reqData.getPassword());
-
-            currentUser = this.userRepository.save(currentUser);
+        if (existing == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User không tồn tại !!!!!");
 
         }
-        return currentUser;
+
+        existing.setGender(GenderEnum.valueOf(reqData.getGender()));
+        existing.setName(reqData.getName());
+        existing.setAddress(reqData.getAddress());
+        existing.setAge(Integer.parseInt(reqData.getAge()));
+
+        existing = this.userRepository.save(existing);
+
+        UserResponseDTO resp = userMapper.toUserResponseDTO(existing);
+        resp.setEmail(null);
+        resp.setCreatedAt(null);
+        return resp;
     }
 
-    public void handleDeleteUser(long id) {
 
+    public void handleDeleteUser(Long id) {
+        User existing = this.handleFindUserById(id);
+
+        if (existing == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User không tồn tại !!!!!");
+
+        }
         this.userRepository.deleteById(id);
     }
 
-    public User handleFindUserByEmail(String email) {
-        return this.userRepository.findUserByEmail(email);
-    }
 }
